LibC: Made mbtowc return int instead of size_t

+++ Libraries/LibC/stdlib.h
@@ -74,7 +74,7 @@ int mkstemp(char*);
 char* mkdtemp(char*);
 void* bsearch(const void* key, const void* base, size_t nmemb, size_t size, int (*compar)(const void*, const void*));
 size_t mbstowcs(wchar_t*, const char*, size_t);
-size_t mbtowc(wchar_t*, const char*, size_t);
+int mbtowc(wchar_t*, const char*, size_t);
 int wctomb(char*, wchar_t);
 size_t wcstombs(char*, const wchar_t*, size_t);
 char* realpath(const char* pathname, char* buffer);
